+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.3.1]
 ### Fix
 - [Mirrorer] not working in maya 2022
+- Python 3 compatability
 
 ## [2.3.0] - 2023-04-26
 ### New